Following recommendations:
* Moved extension to its own namespace and directory
* Added Yard documentation
* Moved tests
