+++ test/test_xml.rb
@@ -102,33 +102,4 @@ class XmlTest < MiniTest::Test
       xml.render.gsub('<', '\n<')
     )
   end
-
-  def test_soap_extension
-    Papercraft.extension(soap: Papercraft::SoapComponents)
-    xml = Papercraft.xml {
-      soap.Envelope(
-        xmlns__xsd: 'http://www.w3.org/2001/XMLSchema',
-        xmlns__xsi: 'http://www.w3.org/2001/XMLSchema-instance'
-      ) {
-        soap.Body {
-          PosRequest(xmlns: 'http://Some.Site') {
-            tag('Ver1.0') {
-              Header {
-                SecretAPIKey 'some_secret_key'
-              }
-              Transaction {
-                SomeData {}
-              }
-            }
-          }
-        }
-      }
-    }
-
-    puts "\n\n[#{xml.render}]\n\n"
-    assert_equal(
-      '<soap:Envelope xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/"><soap:Body><PosRequest xmlns="http://Some.Site"><Ver1.0><Header><SecretAPIKey>some_secret_key</SecretAPIKey></Header><Transaction><SomeData></SomeData></Transaction></Ver1.0></PosRequest></soap:Body></soap:Envelope>'.gsub('<', '\n<'),
-      xml.render.gsub('<', '\n<')
-    )
-  end
 end
